+++ rules/strict-mode.js
@@ -1,0 +1,20 @@
+/**
+ * rules/strict-mode.js: rules listed as Strict Mode in ESLint docs
+ *
+ * Copyright 2017 Sudaraka Wijesinghe <sudaraka@sudaraka.org>
+ *
+ * This program comes with ABSOLUTELY NO WARRANTY;
+ * This is free software, and you are welcome to redistribute it and/or modify
+ * it under the terms of the BSD 2-clause License. See the LICENSE file for more
+ * details.
+ *
+ */
+
+// See https://eslint.org/docs/rules/#strict-mode
+
+module.exports = {
+  'strict': [
+    'error',
+    'safe'
+  ]
+}