+++ rules/stylistic-issues-a2m.js
@@ -1,0 +1,170 @@
+/**
+ * rules/stylistic-issues-a2m.js: rules listed as Stylistic Issues in ESLint docs
+ *
+ * Copyright 2017 Sudaraka Wijesinghe <sudaraka@sudaraka.org>
+ *
+ * This program comes with ABSOLUTELY NO WARRANTY;
+ * This is free software, and you are welcome to redistribute it and/or modify
+ * it under the terms of the BSD 2-clause License. See the LICENSE file for more
+ * details.
+ *
+ */
+
+// NOTE: this file only contain rules starting from A to M
+// See https://eslint.org/docs/rules/#stylistic-issues
+
+module.exports = {
+  'array-bracket-newline': [
+    'error',
+    { 'minItems': 2 }
+  ],
+  'array-bracket-spacing': [
+    'error',
+    'always'
+  ],
+  'array-element-newline': [
+    'error',
+    { 'minItems': 2 }
+  ],
+  'block-spacing': [
+    'error',
+    'always'
+  ],
+  'brace-style': [
+    'error',
+    'stroustrup'
+  ],
+  'camelcase': 'warn',
+  'capitalized-comments': 'off',
+  'comma-dangle': [
+    'error',
+    'never'
+  ],
+  'comma-spacing': [
+    'error',
+    {
+      'before': false,
+      'after': true
+    }
+  ],
+  'comma-style': [
+    'error',
+    'last'
+  ],
+  'computed-property-spacing': [
+    'error',
+    'never'
+  ],
+  'consistent-this': [
+    'error',
+    'self'
+  ],
+  'eol-last': 'error',
+  'func-call-spacing': [
+    'error',
+    'never'
+  ],
+  'func-name-matching': 'error',
+  'func-names': 'off',
+  'func-style': [
+    'error',
+    'expression'
+  ],
+  'id-blacklist': 'off',
+  'id-length': 'off',
+  'id-match': [
+    'error',
+    '^[a-zA-Z0-9_$]+([a-z0-9_]+)*$',
+    { 'properties': true }
+  ],
+  'indent': [
+    'error',
+    2
+  ],
+  'indent-legacy': 'off',
+  'jsx-quotes': [
+    'error',
+    'prefer-single'
+  ],
+  'key-spacing': [
+    'error',
+    {
+      'beforeColon': false,
+      'afterColon': true
+    }
+  ],
+  'keyword-spacing': [
+    'error',
+    {
+      'after': false,
+      'before': true,
+      'overrides': {
+        'else': { 'after': true },
+        'export': { 'after': true },
+        'from': { 'after': true },
+        'import': { 'after': true },
+        'return': { 'after': true },
+        'try': { 'after': true }
+      }
+    }
+  ],
+  'line-comment-position': [
+    'error',
+    { 'position': 'above' }
+  ],
+  'linebreak-style': [
+    'error',
+    'unix'
+  ],
+  'lines-around-comment': [
+    'error',
+    {
+      'allowBlockStart': true,
+      'beforeBlockComment': true,
+      'beforeLineComment': false
+    }
+  ],
+  'lines-around-directive': [
+    'error',
+    {
+      'before': 'never',
+      'after': 'always'
+    }
+  ],
+  'max-depth': [
+    'error',
+    8
+  ],
+  'max-len': [
+    'error',
+    {
+      // 80 should be soft limit for code, using 100 here allow some flexibility
+      'code': 100,
+      'comments': 120,
+      'ignoreComments': true,
+      'ignoreTrailingComments': true,
+      'ignoreUrls': true,
+      'ignoreStrings': true,
+      'ignoreTemplateLiterals': true,
+      'tabWidth': 2
+    }
+  ],
+  'max-lines': 'error',
+  'max-nested-callbacks': [
+    'error',
+    5
+  ],
+  'max-params': [
+    'error',
+    5
+  ],
+  'max-statements': [
+    'error',
+    15
+  ],
+  'max-statements-per-line': [
+    'error',
+    { 'max': 1 }
+  ],
+  'multiline-ternary': 'off'
+}